% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchesWrapper.R
\name{batchesWrapper}
\alias{batchesWrapper}
\title{Batches Wrapper}
\usage{
batchesWrapper(pathfrom, hit_setting_id, question, timed = TRUE,
  pathto = NULL, what = "character", sep = "\\n", quiet = TRUE,
  index = NULL, which_source = "apiR", number_per = 20,
  per_batch = 1000, path = NULL, name = NULL, time_per = 1,
  mintime = 9, maxtime = 22, certone = NULL, certtwo = NULL,
  checkWorkersAt = NULL, rest_time = 60, rate = 1/3, threshold = 5,
  hierarchy_data = NULL, hierarchy_var = NULL, returnFit = FALSE,
  plot = FALSE, file = NULL, chains = 3, iter = 2500, seed = 1234,
  n.cores = 3, ...)
}
\arguments{
\item{hit_setting_id}{ID of HIT setting to use}

\item{question}{the question the worker will see once the worker selects the HIT}

\item{timed}{If True, HITS will be created by time, if not by batch}

\item{what}{The text to be sent and used in the data frame.}

\item{sep}{Where to separate text by line.}

\item{quiet}{If true, this does not print the amount of items read prior.}

\item{index}{The index number}

\item{which_source}{What type of file is the text being drawn from.}

\item{number_per}{number of comparisons desired}

\item{per_batch}{number of comparisons perbatch desired}

\item{path}{File path: if NULL, file will be stored in working direcotry}

\item{name}{File name: If NULL, file will be named pairwise.Rdata}

\item{time_per}{Time per batch to post on to Mechanical Turk}

\item{mintime}{The earliest time in the morning to post comparisons to workers}

\item{maxtime}{The latest time at night to post comparisons to workers}

\item{certone}{Certification to give}

\item{certtwo}{Certification to revoke}

\item{checkWorkersAt}{batch positions to check workers(i.e. batches 1, 3, 5)}

\item{hierarchy_data}{A file that contains the variable that is used as a hierarchy (defalt is NULL)}

\item{hierarchy_var}{A name of the variable in \code{hierarchy_data} that is used as a hierarchy (defalt is NULL)}

\item{returnFit}{Return a fit object if TRUE (default is FALSE)}

\item{plot}{If TRUE, create a histogram with a rug plot (default is FALSE)}

\item{file}{Save the histogram to path and file name specified (default is NULL)}

\item{chains}{The number of chains (default is 3)}

\item{iter}{The number of iteration (default is 2500)}

\item{seed}{Set seed (default is 1234)}

\item{n.cores}{Number of cores to be used in stan fit (default is 3)}

\item{pathFrom}{What file path the data will be drawn form, or actual data}

\item{pathTo}{Where to send the text to be reviewed to.}
}
\value{
batches A vector of batch numbers that have been created
}
\description{
A wrapper function of createBatches, readText
}
\author{
Jacob M. Montgomery
}
\seealso{
\code{\link{batchStatus}}, \code{\link{createHITSTimed}}, \code{\link{checkCert}},\code{\link{checkWorkers}},\code{\link{createBatches}},
\code{\link{createCert}},\code{\link{createHITS}}, \code{\link{createHITSBatch}},\code{\link{createPairwise}}, \code{\link{timedWrapper}},
\code{\link{extractCoef}},\code{\link{fitStan}},\code{\link{fitStanHier}},\code{\link{givetakeCert}},\code{\link{makeCompsSep}},
\code{\link{readInData}}, \code{\link{readText}},\code{\link{repostExpired}},\code{\link{revokeCert}},\code{\link{stanWrapper}}
}

