% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_workers.R
\name{check_workers}
\alias{check_workers}
\title{Detect Outlying Workers}
\usage{
check_workers(fit, data, plot = FALSE, file = NULL)
}
\arguments{
\item{fit}{A stan fit}

\item{data}{The data used to fit stan}

\item{plot}{If TRUE, create a histogram with a rug plot (defalt is FALSE)}

\item{file}{Save the histogram to path and file name specified (defalt is NULL)}
}
\value{
Outlying workers' IDs who have beta coefficients less than 1 and answered more than 100
}
\description{
Detect outlying workers
}
\author{
David Carlson
}
\seealso{
\code{\link{fit_stan}}, \code{\link{fit_stan_hier}}, \code{\link{stanWrapper}}
}

