% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkWorkers.R
\name{checkWorkers}
\alias{checkWorkers}
\title{Detect Outlying Workers}
\usage{
checkWorkers(fit, data, plot = FALSE, file = NULL)
}
\arguments{
\item{fit}{A stan fit}

\item{data}{The data used to fit stan}

\item{plot}{If TRUE, create a histogram with a rug plot (defalt is FALSE)}

\item{file}{Save the histogram to path and file name specified (defalt is NULL)}
}
\value{
Outlying workers' IDs who have beta coefficients less than 1 and answered more than 100
}
\description{
Detect outlying workers
}
\author{
David Carlson
}
\seealso{
\code{\link{batchStatus}}, \code{\link{batchesWrapper}}, \code{\link{createHITSTimed}},\code{\link{checkWorkers}},\code{\link{createBatches}},
\code{\link{createCert}},\code{\link{createHITS}}, \code{\link{createHITSBatch}},\code{\link{createPairwise}}, \code{\link{timedWrapper}},
\code{\link{extractCoef}},\code{\link{fitStan}},\code{\link{fitStanHier}},\code{\link{givetakeCert}},\code{\link{makeCompsSep}},
\code{\link{readInData}}, \code{\link{readText}},\code{\link{repostExpired}},\code{\link{revokeCert}},\code{\link{stanWrapper}}
}

