% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBatches.R
\name{createBatches}
\alias{createBatches}
\title{Create new batches}
\usage{
createBatches(email, password, task_setting_id, num_batches = 1)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{task_setting_id}{ID of Task setting to use}

\item{num_batches}{Number of separate batches to create. Default is 1.}
}
\value{
batch_ids Vector of batch IDs created
}
\description{
Create new batches
}
\details{
Before using this function, you must have created an Amazon Web Service and Mechanical Turk account. You
also must have created an account at SentimentIt.com. This function will not work otherwise. You also should
have run the readText() function to upload your data objects to the SentimentIt server.

The createBatches() function creates unique ID numbers for the batches you will use to submit comparisons
to Mechanical Turk. You must specify how many batches you would like to release your comparisons in. The
function will return a vector of the unique batch IDs you created. This vector should be saved in your
R workspace.

Reference Paper: Carlson, David and Jacob M. Montgomery. Forthcoming. “A Pairwise Comparison Framework for 
Fast, Flexible, and Reliable Human Coding of Political Texts.” American Political Science Review.
}
\examples{
\dontrun{
batch_ids <- createBatches(email = 'researcher@school.edu', password = 'uniquePassword', task_setting_id = 8, num_batches = 3)
}
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
David Carlson
}
