% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCert.R
\name{createCert}
\alias{createCert}
\title{Grants certifications to workers}
\usage{
createCert(email, password, cert, workers)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{cert}{The name of the certification given to the workers.}

\item{workers}{The workers you want to grant certification.}
}
\value{
Character indicating number of workers certified.
}
\description{
Grants certifications to workers
}
\details{
When running an analysis, certifications can be used to control which workers are allowed to access your 
tasks. These certifications allow you to ensure that all workers have passed a training module, and no 
workers that you have determined delivery poor quality work are allowed to continue responding to your 
task. Additional certification documentation can be found here: 
https://www.sentimentit.com/documentation/certifications.markdown.

The createCert() function can be used to assign a certification to workers. If you attempt to assign a 
non-exist certification to a worker, this function will create the certification before assigning it. 
You can require workers to possess a particular certification to work on tasks you send to Mechanical Turk.
}
\examples{

\dontrun{ 
createCert(email = 'researcher@school.edu', password = 'uniquePassword', cert= 'bannedmovie_reviews', workers = ban_workers)
}
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
David Carlson
}
