% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHITS.R
\name{create_hits}
\alias{create_hits}
\title{Create New Hits for comparisons}
\usage{
create_hits(ids = NULL, HITsetting = NULL, batch_id)
}
\arguments{
\item{ids}{Comma-separated list of comparison ids for which to create hits}

\item{HITsetting}{id of the HIT Setting to use for the created HITs}

\item{batch_id}{id of the Batch of comparisons}
}
\value{
out ID for batch of comparisons
}
\description{
This function takes in id numbers of texts and hit settings, or a batch id, and it creates 
HITS for comparison.
}
\note{
This function requires the usage of the httr and jsonlite packages.
}
\author{
David Carlson
}

