% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHITSSmall.R
\name{createHITSSmall}
\alias{createHITSSmall}
\title{Views the hits and data of reviews of two documents.}
\usage{
createHITSSmall(comp_ids = NULL, HITsetting = 9, batch_id)
}
\arguments{
\item{comp_ids}{The id numbers of the texts you want to use.}

\item{batch_id}{Which batch will the reviews be taken from.}
}
\value{
Creates a document with the texts to be compared.
}
\description{
Views the hits and data of reviews of two documents.
}
\note{
This function requires the usage of the httr and jsonlite packages.
}
\author{
David Carlson
}

