% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperTimed.R, R/createHITSTimed.R
\name{wrappertimed}
\alias{wrappertimed}
\title{Wrapper Timed}
\usage{
wrappertimed(batches, time_per, mintime, maxtime, checkWorkersAt = NULL)

wrappertimed(batches, time_per, mintime, maxtime, checkWorkersAt = NULL)
}
\arguments{
\item{batches}{Vector of Batch numbers to check}

\item{batches}{Vector of Batch numbers to check}
}
\value{
out ID for batch of comparisons

out ID for batch of comparisons
}
\description{
Creates batches, and waits a certain

Creates batches, and waits a certain
}
\examples{


}
\author{
Jacob M. Montgomery

Jacob M. Montgomery
}

