% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPairwise.R
\name{createPairwise}
\alias{createPairwise}
\title{Creates a matrix of random pairs for comparison}
\usage{
createPairwise(ids, number_per, pairwise_path = "pairwise.Rdata")
}
\arguments{
\item{ids}{The id numbers of the texts you want to use.}

\item{number_per}{How many documents per batch to be compared.}

\item{pairwise_path}{Path to save the created pairwise matrix. Default is pairwise.Rdata}
}
\value{
A matrix of random pairwise comparisons without repitition.
}
\description{
Creates a matrix of random pairs for comparison
}
\details{
The createPairwise() function randomly selects pairs of ID numbers for objects to be compared. 
This function will return a matrix with two columns, each containing the 

This function randomly samples from all possible pairings of IDs. In other words, no comparison will be duplicated.
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
David Carlson
}
