% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTasks.R
\name{createTasks}
\alias{createTasks}
\title{Create new tasks for comparisons}
\usage{
createTasks(email, password, comp_ids = NULL, task_setting_id = NULL,
  batch_id = NULL)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{comp_ids}{Vector of comparison ids for which to create tasks, if comparisons were not created under a batch. Default is NULL. Leave as NULL if batch_id is provided.}

\item{task_setting_id}{Task setting to use for the created tasks, if comparisons were not created under a batch. Default is NULL. Leave as NULL if batch_id is provided.}

\item{batch_id}{Batch ID to be used if comparisons created under a batch. Default is NULL.}
}
\value{
out ID for batch of comparisons
}
\description{
Create new tasks for comparisons
}
\details{
Before using this function, you must have created an Amazon Web Service and Mechanical Turk account. You
also must have created an account at SentimentIt.com. This function will not work otherwise. You also should
have run the readText() function to upload your data objects to the SentimentIt server. You should have also
run the createBatches() function and saved the batch ID numbers in your R workspace. You should have also
run the makeComps() function to save a portion of your total comparisons under each batch.

To send batches to Mechanical Turk, use the `createTasks()` function. As you can see in the below example, 
you can determine which batches you would like to send to Mechanical Turk by only specifying those specific 
batch IDs. You do not need to send them all at once. 

Note that when you run this function, workers will automatically be paid for the work they do in that batch, 
regardless of the quality of their work. In the SentimentIt system, any worker who completes a task is 
automatically paid every 24 hours.

Reference Paper: Carlson, David and Jacob M. Montgomery. Forthcoming. “A Pairwise Comparison Framework for 
Fast, Flexible, and Reliable Human Coding of Political Texts.” American Political Science Review.
}
\examples{

\dontrun{
createTasks(email = 'researcher@school.edu', password = 'uniquePassword', batch_id = batch_ids[1])
}
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
David Carlson
}
