% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCoef.R
\name{extractCoef}
\alias{extractCoef}
\title{Extract Coefficients}
\usage{
extractCoef(fit, hierarchy_data, document_var, hierarchy_var)
}
\arguments{
\item{fit}{A stan hierarchical model fit}

\item{hierarchy_data}{A file that is used to fit a hierarchical model}

\item{document_var}{A name of the variable in \code{hierarchy_data} that contains document ID number}

\item{hierarchy_var}{A name of the variable in \code{hierarchy_data} that contains a hierarchy's ID number}
}
\description{
Extract coefficients from STAN hierarchical model
}
\author{
David Carlson
}
\seealso{
\code{\link{fitStanHier}}
}

