% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan_hier.R
\name{fit_stan_hier}
\alias{fit_stan_hier}
\title{Fit STAN Hierarchical Model}
\usage{
fit_stan_hier(data, hier_data, chains, iter, seed)
}
\arguments{
\item{data}{A csv file or a vector of batch numbers}

\item{hier_data}{The data used for a hierarchy. One column should be document id and the other column should be a group used as a hierarchy}

\item{chains}{The number of chains (defalt is 3)}

\item{iter}{The number of iteration (defalt is 2500)}

\item{seed}{Set seed (defalt is 1234)}
}
\description{
Fit STAN hierarchical model
}
\author{
David Carlson
}
\seealso{
\code{\link{fit_stan_hier}}, \code{\link{ckeck_workers}}, \code{\link{stanWrapper}}
}

