% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan_square.R
\name{fit_stan_square}
\alias{fit_stan_square}
\title{Fit STAN Model with Square}
\usage{
fit_stan_square(data)
}
\arguments{
\item{data}{a dataset}
}
\value{
STAN output
}
\description{
Fit STAN model with square
}
\author{
David Carlson
}

