% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanRightsDocs}
\alias{humanRightsDocs}
\title{Human rights text data with country IDs and document IDs}
\format{An object of class \code{data.frame} with 1652 rows and 4 columns.}
\usage{
humanRightsDocs
}
\description{
The data set including country torture text, created once the documents were uploaded to the server. Used in fitStanHier() example.
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{revokeCert}} \code{\link{signout}}
}
\keyword{datasets}
