% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeComps_sep.R
\name{makeComps_sep}
\alias{makeComps_sep}
\title{Reads in text.}
\usage{
makeComps_sep(ids, number_per, batches, question, per_batch = 1000,
  path = NULL, name = NULL, idsAsComps = FALSE)
}
\arguments{
\item{ids}{The id numbers of the texts you want to use.}

\item{number_per}{How many documents per batch to be compared.}

\item{batches}{The number of batches to be made.}

\item{question}{Where to separate text by line.}

\item{per_batch}{If true, this does not print the amount of items read prior.}

\item{path}{File path}

\item{name}{File name}

\item{idsAsComps}{IDs as comparison}
}
\value{
A table with the text and correspondings ID's that can be sent.
}
\description{
Reads in text from website to be reviewed.
}
\note{
Makes use of the createPairwise function. Also requires the jsonlite and httr packages.
}
\author{
David Carlson
}

