% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movieReviewOutput}
\alias{movieReviewOutput}
\title{Movie reviews example comparison output from server}
\format{A data frame with 19854 rows and 7 columns:
\describe{
  \item{batch\_id}{The batch ID number from the SentimentIt server}
  \item{comparison\_id}{The comparison ID number from the SentimentIt server for the unique comparison}
  \item{document\_id}{The SentimentIt ID number for the document in the comparison}
  \item{result}{This will be a 1 if this document was chosen in the comparison, and 0 otherwise}
  \item{hit\_id}{This is the ID used by Amazon identifying the HIT, which should not be needed by the researcher}
  \item{worker_id}{This is the ID of the MTurk worker who completed the HIT. THis is used in the package to identify and assess workers}
  \item{completed\_at}{This is the time marker for the time the HIT was competed on MTurk. This can be used to assess the speed MTurkers are completing the HITs}
}}
\usage{
movieReviewOutput
}
\description{
An example output of comparisons created from the SentimentIt system from the movie reviews data set
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{revokeCert}} \code{\link{signout}}
}
\keyword{datasets}
