% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInData.R
\name{readInData}
\alias{readInData}
\title{Read in data provided by the workers}
\usage{
readInData(email, password, batch_id)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{batch_id}{A vector of batch numbers to download data from.}
}
\value{
output a dataframe with the data from the specified batches with columns
\describe{
  \item{batch_id}{the batch number}
  \item{comparison_id}{id number of the comparison being made}
  \item{document_id}{id number of the document}
  \item{result}{result of document comparison}
  \item{task_id}{id of the task}
  \item{worker_id}{id of the worker who did the comparison}
  \item{completed_at}{time comaparison was completed}
}
}
\description{
Read in data provided by the workers
}
\details{
Before using this function, you must have created an Amazon Web Service and Mechanical Turk account. You
also must have created an account at SentimentIt.com. This function will not work otherwise. You also should
have run the readText() function to upload your data objects to the SentimentIt server. You should have also
run the createBatches() function and saved the batch ID numbers in your R workspace. You should have also
run the makeComps() function to save a portion of your total comparisons under each batch. You should have
also run the createTasks() function to send batches of paired data objects to Mechanical Turk for comparison.

Once you have sent the batches to Mechanical Turk using `createTasks()` and allowed some time for comparisons 
to be completed, you can retrieve data on the batches you sent. To retrieve this data, use the `readInData()` 
function. 

You must specify a batch ID to the function, and the function will return a data frame containing the batch ID, 
comparison ID, document ID, result of the comparison, ID of the task, ID of the worker who completed the task, 
and the time when the task was completed.

Reference Paper: Carlson, David and Jacob M. Montgomery. Forthcoming. “A Pairwise Comparison Framework for 
Fast, Flexible, and Reliable Human Coding of Political Texts.” American Political Science Review.
}
\examples{

\dontrun{
output <- readInData(email = 'researcher@school.edu', password = 'uniquePassword', batch_id = batch_ids[1])
}
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
Jacob M. Montgomery
}
