% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentIt-package.R
\docType{package}
\name{sentimentIt}
\alias{sentimentIt}
\alias{sentimentIt-package}
\title{sentimentIt}
\description{
The sentimentIt package performs functions for the sentimentIt project.
}
\examples{


}
\author{
David Carlson
}

