% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanWrapper.R
\name{stanWrapper}
\alias{stanWrapper}
\title{Fit STAN Model and Detect Outlying Workers}
\usage{
stanWrapper(data, hierarchy_data = NULL, hierarchy_var = NULL,
  returnFit = FALSE, plot = FALSE, file = NULL, chains = 3,
  iter = 2500, seed = 1234)
}
\arguments{
\item{data}{A csv file or a vector of batch numbers}

\item{hierarchy_data}{A file that contains the variable that is used as a hierarchy (defalt is NULL)}

\item{hierarchy_var}{A name of the variable in \code{hierarchy_data} that is used as a hierarchy (defalt is NULL)}

\item{returnFit}{Return a fit object if TRUE (degfalt is FALSE)}

\item{plot}{If TRUE, create a histogram with a rug plot (defalt is FALSE)}

\item{file}{Save the histogram to path and file name specified (defalt is NULL)}

\item{chains}{The number of chains (defalt is 3)}

\item{iter}{The number of iteration (defalt is 2500)}

\item{seed}{Set seed (defalt is 1234)}
}
\value{
A list containing
 \item{outlying_workers}{Outlying workers' IDs who have beta coefficients less than 1 and answered more than 100}
 \item{stan_fit}{STAN output}
}
\description{
Fit STAN model or detect outlying workers
}
\author{
David Carlson
}
\seealso{
\code{\link{batchStatus}}, \code{\link{batchesWrapper}}, \code{\link{createHITSTimed}},\code{\link{checkWorkers}},\code{\link{createBatches}},
\code{\link{createCert}},\code{\link{createHITS}}, \code{\link{createHITSBatch}},\code{\link{createPairwise}}, \code{\link{timedWrapper}},
\code{\link{extractCoef}},\code{\link{fitStan}},\code{\link{fitStanHier}},\code{\link{givetakeCert}},\code{\link{makeCompsSep}},
\code{\link{readInData}}, \code{\link{readText}},\code{\link{repostExpired}},\code{\link{revokeCert}},\code{\link{stanWrapper}}
}

