% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchesWrapper.R
\name{batchesWrapper}
\alias{batchesWrapper}
\title{Batches Wrapper}
\usage{
batchesWrapper(hit_setting_id, num_batches = 1, pathfrom, pathto = NULL,
  what = "character", sep = "\\n", quiet = TRUE, index = NULL,
  which_source = "apiR", number_per, batches, question, per_batch = 1000,
  path = NULL, name = NULL, idsAsComps = FALSE, time_per = 1,
  mintime = 8, maxtime = 22, certone = NULL, certtwo = NULL,
  checkWorkersAt = NULL, rest_time = 60, ...)
}
\arguments{
\item{hit_setting_id}{ID of HIT setting to use}

\item{num_batches}{number of batches to create using the HIT setting}

\item{what}{The text to be sent and used in the data frame}

\item{sep}{Where to separate text by line}

\item{quiet}{If true, this does not print the amount of items read prior}

\item{index}{The index number}

\item{which_source}{What type of file is the text being drawn from}

\item{number_per}{How many documents per batch to be compared}

\item{batches}{The number of batches to be made}

\item{question}{Where to separate text by line}

\item{per_batch}{If true, this does not print the amount of items read prior}

\item{path}{File path}

\item{name}{File name}

\item{idsAsComps}{IDs as comparison}

\item{time_per}{Defalt is 1}

\item{mintime}{Defalt is 8}

\item{maxtime}{Defalt is 22}

\item{certone}{Defalt is NULL}

\item{certtwo}{Defalt isNULL}

\item{checkWorkersAt}{Defalt is NULL}

\item{rest_time}{Defalt is 60}

\item{pathFrom}{Where the text will be drawn from}

\item{pathTo}{Where to send the text to be reviewed to}
}
\value{
batches A vector of batch numbers
}
\description{
A wrapper function of createBatches, readText
}
\examples{

}
\author{
Jacob M. Montgomery
}
\seealso{
\code{\link{createBatches}}, \code{\link{readText}}, \code{\link{makeCompsSep}}, \code{\link{createHITSTimed}}
}

