% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchStatus.R
\name{batchStatus}
\alias{batchStatus}
\title{Check the status of batch(es)}
\usage{
batchStatus(email, password, batch_id)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{batch_id}{Vector of batch IDs to check}
}
\value{
output Dataframe with the variables:
\itemize{
  \item id
  \item total_count
  \item submitted count
  \item completed_count
  \item expired_count
}
}
\description{
Check the status of batch(es)
}
\details{
Before using this function, you must have created an Amazon Web Service and Mechanical Turk account. You
also must have created an account at SentimentIt.com. This function will not work otherwise. You also should
have run the readText() function to upload your data objects to the SentimentIt server. You should have also
run the createBatches() function and saved the batch ID numbers in your R workspace. You should have also
run the makeComps() function to save a portion of your total comparisons under each batch. You should have
also run the createTasks() function to send batches of paired data objects to Mechanical Turk for comparison.

Once you have submitted batches of comparisons to Mechanical Turk, you can check their status to see how many 
comparisons in the batch have been completed. This can be done using the `checkBatches()` function. To check 
the status of a batch, all you need is the batch ID number.

Reference Paper: Carlson, David and Jacob M. Montgomery. Forthcoming. “A Pairwise Comparison Framework for 
Fast, Flexible, and Reliable Human Coding of Political Texts.” American Political Science Review.
}
\examples{

\dontrun{ 
batchStatus(email = 'researcher@school.edu', password = 'uniquePassword', batch_id = batch_ids[1])
}
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
David Carlson
}
