% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchStatus.R
\name{batchStatus}
\alias{batchStatus}
\title{Check batch status}
\usage{
batchStatus(email, password, batch_id)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{batch_id}{Vector of batch IDs to check}
}
\value{
output Dataframe with the variables:
\itemize{
  \item id
  \item total_count
  \item submitted count
  \item completed_count
  \item expired_count
}
}
\description{
This function checks the indicated batch_id numbers and returns their status. This returns a data frame with the batch ID and the number of comparisons submitted and the number completed.
}
\examples{

\dontrun{ 
batch_id <- 204:208
status <- batchStatus(email, password, batch_id)
}
}
\author{
David Carlson
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeCompsSep}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{revokeCert}} \code{\link{signout}}
}

