% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchStatus.R
\name{batch_status}
\alias{batch_status}
\title{Check batch status}
\usage{
batch_status(batch_id)
}
\arguments{
\item{batch_id}{ID of batch to check}
}
\value{
output dataframe with the variables:
\itemize{
  \item id
  \item total_count
  \item submitted count
  \item completed_count
  \item expired_count
}
}
\description{
This function checks the indicated batch_id numbers and returns their status.
}
\note{
This function requires the usage of the jsonlite packages
}
\examples{

x <- 204:208
status <- batchStatus(x)
y <- c(204, 206, 208, 207)
status2 <- batchStatus(y)
}
\author{
David Carlson
}

