% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCert.R
\name{checkCert}
\alias{checkCert}
\title{Checks if a worker has a certification}
\usage{
checkCert(email, password, cert, worker)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{cert}{The name of the certification given to the workers.}

\item{worker}{The MTurk worker ID you want to check for certification.}
}
\value{
TRUE or FALSE indicating if the worker has the certification.
}
\description{
Checks if a worker has a certification
}
\details{
When running an analysis, certifications can be used to control which workers are allowed to access your 
tasks. These certifications allow you to ensure that all workers have passed a training module, and no 
workers that you have determined delivery poor quality work are allowed to continue responding to your 
task. Additional certification documentation can be found here: 
https://www.sentimentit.com/documentation/certifications.markdown.

To check if a worker has a particular certification, use the checkCert() function. To use this function, 
you must know the name of the certification you would like to check for as well as the Mechanical Turk ID 
for the worker of interest.
}
\examples{

\dontrun{ 
x <- "ab1"
y <- c("a204")
check <- checkCert(email, password, x, y)
}

}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
David Carlson
}
