% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCert.R
\name{checkCert}
\alias{checkCert}
\title{Checks if a worker has a certification}
\usage{
checkCert(email, password, cert, worker)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{cert}{The name of the certification given to the workers.}

\item{worker}{The MTurk worker ID you want to check for certification.}
}
\value{
TRUE or FALSE indicating if the worker has the certification.
}
\description{
Checks if a worker has a certification
}
\examples{

\dontrun{ 
x <- "ab1"
y <- c("a204")
check <- checkCert(email, password, x, y)
}

}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
David Carlson
}
