% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkWorkers.R, R/checkWorkers00.R
\name{checkWorkers}
\alias{checkWorkers}
\title{Detect Outlying Workers}
\usage{
checkWorkers(stan_fit, data, plot_hist = FALSE, file_path = NULL)

checkWorkers(stan_fit, data, plot_hist = FALSE, file_path = NULL)
}
\arguments{
\item{stan_fit}{A stan fit.}

\item{data}{The data used to fit stan.}

\item{plot_hist}{If TRUE, plot the histogram of workers with a rug plot. (Default is FALSE)}

\item{file_path}{Save the histogram to path and file name specified. (Default is NULL)}

\item{cut_point}{A cutoff point to classify posterior coefficients. The proportion of posterior coefficients below \code{cut_point} is used to determine outliers. (Default is 1)}

\item{cut_proportion}{A cutoff proportion of posterior coefficients below \code{cut_point}. If the proportion of posterior coefficients below \code{cut_points} is higher than \code{cut_proportion}, a worker will be considered as an outlier provided that she answers more than 50 questions. (Default is 0.9)}

\item{stan_fit}{A stan fit}

\item{data}{The data used to fit stan}

\item{plot_hist}{If TRUE, plot the histogram of workers with a rug plot (default is FALSE)}

\item{file_path}{Save the histogram to path and file name specified (default is NULL)}
}
\value{
A vector of outlying workers' IDs whose proportion of posterior coefficients below \code{cut_point} is greater than \code{cut_proportion} and who answered more than 50 questions

Outlying workers' IDs who have beta coefficients less than 1 and answered more than 100
}
\description{
Detect outlying workers

Detect outlying workers
}
\author{
David Carlson

David Carlson
}
\seealso{
\code{\link{batchStatus}}, \code{\link{batchesWrapper}}, \code{\link{createHITSTimed}},\code{\link{checkWorkers}},\code{\link{createBatches}},
\code{\link{createCert}},\code{\link{createHITS}}, \code{\link{createHITSBatch}},\code{\link{createPairwise}}, \code{\link{timedWrapper}},
\code{\link{extractCoef}},\code{\link{fitStan}},\code{\link{fitStanHier}},\code{\link{givetakeCert}},\code{\link{makeCompsSep}},
\code{\link{readInData}}, \code{\link{readText}},\code{\link{repostExpired}},\code{\link{revokeCert}},\code{\link{stanWrapper}}

\code{\link{batchStatus}}, \code{\link{batchesWrapper}}, \code{\link{createHITSTimed}},\code{\link{checkWorkers}},\code{\link{createBatches}},
\code{\link{createCert}},\code{\link{createHITS}}, \code{\link{createHITSBatch}},\code{\link{createPairwise}}, \code{\link{timedWrapper}},
\code{\link{extractCoef}},\code{\link{fitStan}},\code{\link{fitStanHier}},\code{\link{givetakeCert}},\code{\link{makeCompsSep}},
\code{\link{readInData}}, \code{\link{readText}},\code{\link{repostExpired}},\code{\link{revokeCert}},\code{\link{stanWrapper}}
}

