% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBatches.R
\name{createBatches}
\alias{createBatches}
\title{Create new batches}
\usage{
createBatches(hit_setting_id, num_batches = 1)
}
\arguments{
\item{num_batches}{number of batches to create using the HIT setting}

\item{ids}{ID of HIT setting to use}
}
\value{
out ID of batches created
}
\description{
This function creates new batches with a HIT setting.
}
\examples{

createHITS(ids=10,HITsetting=2)
createHITS(batch_id=204)
}
\author{
David Carlson
}

