% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCert.R
\name{createCert}
\alias{createCert}
\title{Grants certifications to workers.}
\usage{
createCert(email, password, cert, workers)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{cert}{The name of the certification given to the workers.}

\item{workers}{The workers you want to grant certification.}
}
\value{
Character indicating number of workers certified.
}
\description{
Grants certifications to workers.
}
\examples{

\dontrun{ 
x <- "ab1"
y <- c("a204", "a206", "a208", "a207")
creation <- createCert(email, password, x, y)
}
}
\author{
David Carlson
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeCompsSep}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}

