% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCert.R
\name{createCert}
\alias{createCert}
\title{Grants certifications to workers}
\usage{
createCert(email, password, cert, workers)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{cert}{The name of the certification given to the workers.}

\item{workers}{The workers you want to grant certification.}
}
\value{
Character indicating number of workers certified.
}
\description{
This function will grant a certification to a worker, and if the certification does not exist it will be created.
}
\examples{

\dontrun{ 
createCert(email = 'researcher@school.edu', password = 'uniquePassword', cert= 'bannedmovie_reviews', workers = ban_workers)
}
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
David Carlson
}
