% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHITSTimed.R
\name{createHITStimed}
\alias{createHITStimed}
\title{Create HITS Timed}
\usage{
createHITStimed(batches, time_per, mintime, maxtime, certone, certtwo,
  checkWorkersAt = NULL, hierarchy_data = NULL, hierarchy_var = NULL,
  returnFit = FALSE, plot = FALSE, file = NULL, chains = 3,
  iter = 2500, seed = 1234, n.cores = 3)
}
\arguments{
\item{batches}{Vector of Batch numbers to check}

\item{time_per}{Time per batch to post on to Mechanical Turk}

\item{mintime}{The earliest time in the morning to post comparisons to workers}

\item{maxtime}{The latest time at night to post comparisons to workers}

\item{certone}{The name of the certification wanted for the workers.}

\item{certtwo}{The name of the certification wanted to be removed from the workers.}

\item{checkWorkersAt}{batch positions to check workers(i.e. batches 1, 3, 5)}

\item{hierarchy_data}{A file that contains the variable that is used as a hierarchy (defalt is NULL)}

\item{hierarchy_var}{A name of the variable in \code{hierarchy_data} that is used as a hierarchy (defalt is NULL)}

\item{returnFit}{Return a fit object if TRUE (degfalt is FALSE)}

\item{plot}{If TRUE, create a histogram with a rug plot (defalt is FALSE)}

\item{file}{Save the histogram to path and file name specified (defalt is NULL)}

\item{chains}{The number of chains (defalt is 3)}

\item{iter}{The number of iteration (defalt is 2500)}

\item{seed}{Set seed (defalt is 1234)}

\item{n.cores}{Number of cores to be used in stan fit (default is 3)}

\item{certone}{Certification to give}

\item{certtwo}{Certification to revoke}
}
\value{
out IDs for batches of comparisons
}
\description{
Creates batches, and waits a certain
}
\author{
Jacob M. Montgomery
}
\seealso{
\code{\link{batchStatus}}, \code{\link{batchesWrapper}}, \code{\link{checkCert}},\code{\link{checkWorkers}},\code{\link{createBatches}},
\code{\link{createCert}},\code{\link{createHITS}}, \code{\link{createHITSBatch}},\code{\link{createPairwise}}, \code{\link{timedWrapper}},
\code{\link{extractCoef}},\code{\link{fitStan}},\code{\link{fitStanHier}},\code{\link{givetakeCert}},\code{\link{makeCompsSep}},
\code{\link{readInData}}, \code{\link{readText}},\code{\link{repostExpired}},\code{\link{revokeCert}},\code{\link{stanWrapper}}
}

