% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHITSTimed.R
\name{createHITStimed}
\alias{createHITStimed}
\title{Wrapper Timed}
\usage{
createHITStimed(batches, time_per, mintime, maxtime, checkWorkersAt = NULL)
}
\arguments{
\item{batches}{Vector of Batch numbers to check}
}
\value{
out ID for batch of comparisons
}
\description{
Creates batches, and waits a certain
}
\examples{

}
\author{
Jacob M. Montgomery
}

