% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPairwise.R
\name{createPairwise}
\alias{createPairwise}
\title{Creates a matrix of random pairs for comparison.}
\usage{
createPairwise(ids, number_per, pairwise_path = "pairwise.Rdata")
}
\arguments{
\item{ids}{The id numbers of the texts you want to use.}

\item{number_per}{How many documents per batch to be compared.}

\item{pairwise_path}{Path to save the created pairwise matrix. Default is pairwise.Rdata}
}
\value{
A matrix of random pairwise comparisons without repitition.
}
\description{
Creates a matrix of random pairs for comparison.
}
\author{
David Carlson
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeCompsSep}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}

