% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPairwise.R
\name{createPairwise}
\alias{createPairwise}
\title{Creates documents with two texts to be compared.}
\usage{
createPairwise(ids, number_per)
}
\arguments{
\item{ids}{The id numbers of the texts you want to use.}

\item{number_per}{How many documents per batch to be compared.}
}
\value{
Creates a document with the texts to be compared.
}
\description{
Creates documents with two texts to be compared.
}
\note{
This function requires the usage of the plyr package.
}
\author{
David Carlson
}

