% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTasks.R
\name{createTasks}
\alias{createTasks}
\title{Create new tasks for comparisons}
\usage{
createTasks(email, password, comp_ids = NULL, task_setting_id = NULL,
  batch_id = NULL)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{comp_ids}{Vector of comparison ids for which to create tasks, if comparisons were not created under a batch. Default is NULL. Leave as NULL if batch_id is provided.}

\item{task_setting_id}{Task setting to use for the created tasks, if comparisons were not created under a batch. Default is NULL. Leave as NULL if batch_id is provided.}

\item{batch_id}{Batch ID to be used if comparisons created under a batch. Default is NULL.}
}
\value{
out ID for batch of comparisons
}
\description{
This function posts a batch or comparisons to Mechanical Turk as tasks
}
\examples{

\dontrun{
createTasks(email = 'researcher@school.edu', password = 'uniquePassword', batch_id = batch_ids[1])
}
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
David Carlson
}
