% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan_hier.R, R/fit_stan_square.R
\name{fit_stan_square}
\alias{fit_stan_square}
\title{Fit STAN Hierarchical Model}
\usage{
fit_stan_square(data)

fit_stan_square(data)
}
\arguments{
\item{data}{a dataset}

\item{data}{a dataset}
}
\value{
STAN output

STAN output
}
\description{
Fit STAN hierarchical model

Fit STAN model with square
}
\author{
David Carlson

David Carlson
}

