% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStan.R
\name{fitStan}
\alias{fitStan}
\title{Fit a Stan model with results from SentimentIt}
\usage{
fitStan(email = NULL, password = NULL, data, chains = 3, iter = 2500,
  seed = 1234, n.cores = 3)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration. Default is NULL and only needs to be provided if batch numbers are used instead of data.}

\item{password}{The researcher's password used for SentimentIt. Default is NULL and only needs to be provided if batch numbers are used instead of data.}

\item{data}{A csv file or a vector of batch numbers.}

\item{chains}{The number of chains. (Default is 3)}

\item{iter}{The number of iteration. (Default is 2500)}

\item{seed}{Set seed. (Defalt is 1234)}

\item{n.cores}{Number of cores to be used in stan fit. (Default is 3)}
}
\value{
fit Stan fit object
}
\description{
Fit a random utility model using Hamiltonian MCMC in Stan with data retrieved from the SentimentIt platform.
}
\examples{

\dontrun{
fit <- fitStan(data = batch_ids)
fit <- fitStan(data = output)
}

}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
David Carlson
}
