% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStan.R
\name{fitStan}
\alias{fitStan}
\title{Fit STAN Model}
\usage{
fitStan(data, chains = 3, iter = 2500, seed = 1234)
}
\arguments{
\item{data}{A csv file or a vector of batch numbers}

\item{chains}{The number of chains (defalt is 3)}

\item{iter}{The number of iteration (defalt is 2500)}

\item{seed}{Set seed (defalt is 1234)}
}
\value{
STAN output
}
\description{
Fit STAN model
}
\author{
David Carlson
}
\seealso{
\code{\link{batchStatus}}, \code{\link{batchesWrapper}}, \code{\link{createHITSTimed}},\code{\link{checkWorkers}},\code{\link{createBatches}},
\code{\link{createCert}},\code{\link{createHITS}}, \code{\link{createHITSBatch}},\code{\link{createPairwise}}, \code{\link{timedWrapper}},
\code{\link{extractCoef}},\code{\link{fitStan}},\code{\link{fitStanHier}},\code{\link{givetakeCert}},\code{\link{makeCompsSep}},
\code{\link{readInData}}, \code{\link{readText}},\code{\link{repostExpired}},\code{\link{revokeCert}},\code{\link{stanWrapper}}
}

