% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStanHier.R
\name{fitStanHier}
\alias{fitStanHier}
\title{Fit STAN Hierarchical Model}
\usage{
fitStanHier(data, hierarchy_data, hierarchy_var, chains = 3, iter = 2500,
  seed = 1234)
}
\arguments{
\item{data}{A csv file or a vector of batch numbers}

\item{hierarchy_data}{A file that contains the variable that is used as a hierarchy}

\item{hierarchy_var}{A name of the variable in \code{hierarchy_data} that is used as a hierarchy}

\item{chains}{The number of chains (defalt is 3)}

\item{iter}{The number of iteration (defalt is 2500)}

\item{seed}{Set seed (defalt is 1234)}
}
\value{
fitStanHier
}
\description{
Fit STAN hierarchical model
}
\author{
David Carlson
}
\seealso{
\code{\link{fit_stan_hier}}, \code{\link{ckeck_workers}}, \code{\link{stanWrapper}}
}

