% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStanHier.R
\name{fitStanHier}
\alias{fitStanHier}
\title{Fit STAN Hierarchical Model}
\usage{
fitStanHier(data, hierarchy_data, hierarchy_var, chains = 3, iter = 2500,
  seed = 1234, n.cores = 3)
}
\arguments{
\item{data}{A csv file or a vector of batch numbers.}

\item{hierarchy_data}{A file that contains the variable that is used as a hierarchy.}

\item{hierarchy_var}{A name of the variable in \code{hierarchy_data} that is used as a hierarchy.}

\item{chains}{The number of chains. (Default is 3)}

\item{iter}{The number of iteration. (Default is 2500)}

\item{seed}{Set seed. (Default is 1234)}

\item{n.cores}{Number of cores to be used in stan fit. (Default is 3)}
}
\value{
fitStanHier
}
\description{
Fit STAN hierarchical model
}
\author{
David Carlson
}
\seealso{
\code{\link{batchStatus}}, \code{\link{batchesWrapper}}, \code{\link{createHITSTimed}},\code{\link{checkWorkers}},\code{\link{createBatches}},
\code{\link{createCert}},\code{\link{createHITS}}, \code{\link{createHITSBatch}},\code{\link{createPairwise}}, \code{\link{timedWrapper}},
\code{\link{extractCoef}},\code{\link{fitStan}},\code{\link{fitStanHier}},\code{\link{givetakeCert}},\code{\link{makeCompsSep}},
\code{\link{readInData}}, \code{\link{readText}},\code{\link{repostExpired}},\code{\link{revokeCert}},\code{\link{stanWrapper}}
}

