% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInData.R
\name{readInData}
\alias{readInData}
\title{Read in Data}
\usage{
readInData(batchNumbers)
}
\arguments{
\item{batchNumbers}{a vector of batch numbers to download data from.}
}
\value{
output a dataframe with the data from the specified batches
}
\description{
Reads in data for specified batch numbers from the website
}
\note{
Sometimes the server will fail to send the data to your computer, 
     so this function will try again until it successfully connects to the server.
}
\examples{

x <- 204:208
myData <- readInData(x) 
y <- c(204, 206, 208, 207)
myData2 <- readInData(y)
}
\author{
Jacob M. Montgomery
}

