% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInData.R
\name{readInData}
\alias{readInData}
\title{Read in data provided by the workers}
\usage{
readInData(email, password, batch_id)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{batch_id}{A vector of batch numbers to download data from.}
}
\value{
output a dataframe with the data from the specified batches with columns
\describe{
  \item{batch_id}{the batch number}
  \item{comparison_id}{id number of the comparison being made}
  \item{document_id}{id number of the document}
  \item{result}{result of document comparison}
  \item{task_id}{id of the task}
  \item{worker_id}{id of the worker who did the comparison}
  \item{completed_at}{time comaparison was completed}
}
}
\description{
Reads in the complete data for specified batch numbers from the server. This data contains the selection made for every completed comparison in the batch.
}
\examples{

\dontrun{
output <- readInData(email = 'researcher@school.edu', password = 'uniquePassword', batch_id = batch_ids[1])
}
}
\author{
Jacob M. Montgomery
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeCompsSep}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}

