% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readText.R
\name{readText}
\alias{readText}
\title{Find/create documents and retrieve IDs}
\usage{
readText(email, password, read_documents_from, write_documents_to = NULL,
  what = "character", sep = "\\n", quiet = TRUE, index = NULL,
  which_source = "apiR", ...)
}
\arguments{
\item{email}{The researcher's email associated with the SentimentIt account.}

\item{password}{The researcher's password associated with the SentimentIt account.}

\item{read_documents_from}{The file path the data will be drawn form, or a vector of text.}

\item{write_documents_to}{The file path to write the original data, merged with the document IDs. Default is NULL and the results will not be saved, but only returned. For best functionality specify a csv.}

\item{what}{Argument passed to scan() function. Default is character. Only needed when index is NULL.}

\item{sep}{Argument passed to read.table() function. Default is line break. Only needed when index is not NULL.}

\item{quiet}{Argument passed to scan(). Default is TRUE. Only needed when index is NULL.}

\item{index}{The index number of the table to extract the text from, or the name of the column. Default is NULL, indicating the text was not sent in a table.}

\item{which_source}{Source used within SentimentIt server assoicated with document uploads. Only used for later reference. Default is apiR.}

\item{...}{Additional arguments passed to either scan() or read.table() depending on type of data used}
}
\value{
A data frame with the provided data and the IDs appended
}
\description{
Reads in text, posts the text to the server, and returns the document IDs after saving the result to a new table. If text already exists on the server, the associated ID is returned (there will never be duplicated text).
}
\examples{

\dontrun{
data(reviews)

docInfo <- readText(email = 'researcher@school.edu', password = 'uniquePassword', read_documents_from = reviews, write_documents_to = "ReviewsWithIds.csv", index = 'Review')
}

}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeCompsSep}} \code{\link{readInData}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}
\author{
David Carlson
}
