% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readText.R
\name{readText}
\alias{readText}
\title{Reads in text.}
\usage{
readText(pathfrom, pathto = NULL, what = "character", sep = "\\n",
  quiet = TRUE, index = NULL, which_source = "apiR", ...)
}
\arguments{
\item{what}{The text to be sent and used in the data frame.}

\item{sep}{Where to separate text by line.}

\item{quiet}{If true, this does not print the amount of items read prior.}

\item{index}{The index number}

\item{which_source}{What type of file is the text being drawn from.}

\item{pathFrom}{Where the text will be drawn from.}

\item{pathTo}{Where to send the text to be reviewed to.}
}
\value{
A table with the text and correspondings ID's that can be sent.
}
\description{
Reads in text from website to be reviewed.
}
\author{
David Carlson
}

