% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repostExpired.R
\name{repostExpired}
\alias{repostExpired}
\title{Repost expired HITS from a batch}
\usage{
repostExpired(batch_id)
}
\arguments{
\item{batch_id}{ID of batch to check}
}
\value{
out reposted count
}
\description{
This function takes in expired batch_ids and reposts them to create new comparisons.
}
\examples{

repostExpired(batch_id=2)
}
\author{
David Carlson
}

