% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repostExpired.R
\name{repostExpired}
\alias{repostExpired}
\title{Repost expired Tasks from a batch}
\usage{
repostExpired(email, password, batch_id)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{batch_id}{ID of batch to check}
}
\value{
out reposted count
}
\description{
This will repost all of the expired Tasks from the vector of batch IDs, because once all
batches are complete it is common that a few Tasks remain incomplete.
}
\examples{
\dontrun{
repostExpired(email, password, batch_id=2)
}
}
\author{
David Carlson
}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeCompsSep}} \code{\link{readInData}} \code{\link{readText}} \code{\link{reviews}} \code{\link{revokeCert}} \code{\link{signout}}
}

