% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revokeCert.R
\name{revokeCert}
\alias{revokeCert}
\title{Revoke a certification for workers}
\usage{
revokeCert(email, password, cert, workers)
}
\arguments{
\item{email}{The researcher's email used for SentimentIt registration}

\item{password}{The researcher's password used for SentimentIt}

\item{cert}{The name of the certification given to the workers.}

\item{workers}{The workers you want to grant certification.}
}
\value{
The number of workers revoked.
}
\description{
This function simply removes workers from the list of approved workers and they will no longer be able to perform the tasks requiring the certification.
}
\examples{

\dontrun{ 
revokeCert(email = 'researcher@school.edu', password = 'uniquePassword', cert = 'snippets', workers = ban_workers)
}

}
\seealso{
\code{\link{sentimentIt}} \code{\link{authenticate}} \code{\link{batchStatus}} \code{\link{checkCert}} \code{\link{checkWorkers}} \code{\link{createBatches}} \code{\link{createCert}} \code{\link{createPairwise}} \code{\link{createTasks}} \code{\link{fitStan}} \code{\link{fitStanHier}} \code{\link{makeComps}} \code{\link{readInData}} \code{\link{readText}} \code{\link{repostExpired}} \code{\link{reviews}} \code{\link{signout}}
}
\author{
David Carlson
}
