% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanWrapper.R
\name{stanWrapper}
\alias{stanWrapper}
\title{Fit STAN Model and Detect Outlying Workers}
\usage{
stanWrapper(data, hier_data = NULL, returnFit = FALSE, plot = FALSE,
  file = NULL, chains = 3, iter = 2500, seed = 1234)
}
\arguments{
\item{data}{A csv file or a vector of batch numbers}

\item{hier_data}{The data used for a hierarchy. One column should be document id and the other column should be a group used as a hierarchy. Defalt is NULL, and unless speficied, this function fits fit_stan}

\item{returnFit}{Return a fit object if TRUE (degfalt is FALSE)}

\item{plot}{If TRUE, create a histogram with a rug plot (defalt is FALSE)}

\item{file}{Save the histogram to path and file name specified (defalt is NULL)}

\item{chains}{The number of chains (defalt is 3)}

\item{iter}{The number of iteration (defalt is 2500)}

\item{seed}{Set seed (defalt is 1234)}
}
\value{
A list containing
 \item{outlying_workers}{Outlying workers' IDs who have beta coefficients less than 1 and answered more than 100}
 \item{stan_fit}{STAN output}
}
\description{
Fit STAN model or detect outlying workers
}
\author{
David Carlson
}
\seealso{
\code{\link{fit_stan_hier}}, \code{\link{ckeck_workers}}, \code{\link{stanWrapper}}
}

