% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanWrapper.R
\name{.stanWrapper}
\alias{.stanWrapper}
\title{Fit STAN Model and Detect Outlying Workers}
\usage{
.stanWrapper(data, hierarchy_data = NULL, hierarchy_var = NULL,
  returnFit = FALSE, cut_point = 1, cut_proportion = 0.9,
  n.questions = 50, plot_hist = FALSE, file_path = NULL, chains = 3,
  iter = 2500, seed = 1234, n.cores = 3)
}
\arguments{
\item{data}{A csv file or a vector of batch numbers.}

\item{hierarchy_data}{A file that contains the variable that is used as a hierarchy. (Default is NULL)}

\item{hierarchy_var}{A name of the variable in \code{hierarchy_data} that is used as a hierarchy. (Default is NULL)}

\item{returnFit}{Return a fit object if TRUE. (Default is FALSE)}

\item{cut_point}{A cutoff point to classify posterior coefficients. The proportion of posterior coefficients below \code{cut_point} is used to determine outliers. (Default is 1)}

\item{cut_proportion}{A cutoff proportion of posterior coefficients below \code{cut_point}. If the proportion of posterior coefficients below \code{cut_points} is higher than \code{cut_proportion}, a worker will be considered as an outlier provided that she answers more than the number of questions in \code{n.questions}. (Default is 0.9)}

\item{n.questions}{The number of questions to consider in order to determine banned workers. (Default is 50)}

\item{plot_hist}{If TRUE, plot the histogram of workers with a rug plot. (Default is FALSE)}

\item{file_path}{Save the histogram to path and file name specified. (Default is NULL)}

\item{chains}{The number of chains. (Default is 3)}

\item{iter}{The number of iteration. (Default is 2500)}

\item{seed}{Set seed. (Default is 1234)}

\item{n.cores}{Number of cores to be used in stan fit. (Default is 3)}
}
\value{
A list containing
 \item{outlying_workers}{A vector of outlying workers' IDs whose proportion of posterior coefficients below \code{cut_point} is greater than \code{cut_proportion} and who answered more than 50 questions.}
 \item{stan_fit}{STAN output (class stanfit)}
}
\description{
A wrapper function for fitStan, fitStanHier, and checkWorkers. Fit STAN model and/or detect outlying workers.
}
\author{
David Carlson
}
\seealso{
\code{\link{batchStatus}}, \code{\link{batchesWrapper}}, \code{\link{createHITSTimed}},\code{\link{checkWorkers}},\code{\link{createBatches}},
\code{\link{createCert}},\code{\link{createHITS}}, \code{\link{createHITSBatch}},\code{\link{createPairwise}}, \code{\link{timedWrapper}},
\code{\link{extractCoef}},\code{\link{fitStan}},\code{\link{fitStanHier}},\code{\link{givetakeCert}},\code{\link{makeCompsSep}},
\code{\link{readInData}}, \code{\link{readText}},\code{\link{repostExpired}},\code{\link{revokeCert}},\code{\link{stanWrapper}}
}

